+++ addons/pt/saft/extensions.go
@@ -291,11 +291,21 @@ var extensions = []*cbc.Definition{
 		},
 		Desc: i18n.String{
 			i18n.EN: here.Doc(`
-				AT's ~TaxExemptionCode~ (Código do motivo de isenção de imposto) is a code that
+				The SAF-T's ~TaxExemptionCode~ (Código do motivo de isencção de imposto) is a code that
 				specifies the reason the VAT tax is exempt in a Portuguese invoice. When the ~exempt~ tag
-				is used in a tax combo, the ~ext~ map's ~pt-exemption-code~ property is required.
+				is used in a tax combo, the ~ext~ map's ~pt-saft-exemption~ property is required.
 
-				For example, you could define an invoice line exempt of tax as follows:
+				The SAF-T's ~TaxExemptionReason~ (Motivo da isenção de imposto) is a text that justifies
+				the exemption referencing the relevant legislation. In GOBL, this is provided with a
+				special line-level ~note~.
+
+				By default, if no note is provided, GOBL will automatically assign a default one
+				consistent with the exemption code. However, note that default texts are generic and not
+				always sufficiently precise to comply with the regulations. The invoice issuer should be
+				given the option to provide a custom note with the appropriate descriptive text.
+
+				For example, you could define an invoice line exempt of tax with the exemption code and
+				the reason as follows:
 
 				~~~js
 				{
@@ -314,11 +324,19 @@ var extensions = []*cbc.Definition{
 										"rate": "exempt",
 										"ext": {
 											"pt-saft-tax-rate": "ISE",
-											"pt-saft-exemption": "M19"
+											"pt-saft-exemption": "M40"
 										}
 								}
 							]
-						}
+						},
+						"notes": [
+							{
+								"key": "legal",
+								"code": "M40",
+								"src": "pt-saft-exemption",
+								"text": "Artigo 6.º n.º 6 alínea a) do CIVA, a contrário"
+							}
+						]
 					]
 				}
 				~~~
@@ -327,19 +345,38 @@ var extensions = []*cbc.Definition{
 		Sources: []*cbc.Source{
 			{
 				Title: i18n.String{
-					i18n.PT: "AT Tax Exemption Codes",
-					i18n.EN: "Códigos de motivo de isenção",
+					i18n.EN: "AT Tax Exemption Codes",
+					i18n.PT: "Códigos de motivo de isenção",
 				},
 				URL:         "https://info.portaldasfinancas.gov.pt/pt/apoio_contribuinte/Faturacao/Fatcorews/Documents/Tabela_Codigos_Motivo_Isencao.pdf",
 				ContentType: "application/pdf",
 			},
+			{
+				Title: i18n.String{
+					i18n.EN: "Art. 2.2.14 of Despacho nº8632/2014",
+					i18n.PT: "Art. 2.2.14 do Despacho nº8632/2014",
+				},
+				URL:         "https://files.diariodarepublica.pt/2s/2014/07/126000000/1725517261.pdf",
+				ContentType: "application/pdf",
+			},
+			{
+				Title: i18n.String{
+					i18n.EN: "Field 4.4.19.7 of Portaria nº302/2016",
+					i18n.PT: "Campo 4.4.19.7 da Portaria nº302/2016",
+				},
+				URL:         "https://files.diariodarepublica.pt/1s/2016/12/23100/0427304379.pdf",
+				ContentType: "application/pdf",
+			},
 		},
 		Values: []*cbc.Definition{
+			// Names adapted to meet the requirements of art. 2.2.14 of
+			// Despacho nº8632/2014 (for printed invoices) and campo 4.4.19.7
+			// of Portaria nº302/2016 (for SAF-T files).
 			{
 				Code: "M01",
 				Name: i18n.String{
-					i18n.EN: "Article 16, No. 6 of the VAT code",
-					i18n.PT: "Artigo 16.º, n.º 6 do CIVA",
+					i18n.EN: "Article 16, No. 6, paragraphs a) to d) of the VAT code",
+					i18n.PT: "Artigo 16.º, n.º 6, alíneas a) a d) do CIVA",
 				},
 			},
 			{
@@ -352,85 +389,85 @@ var extensions = []*cbc.Definition{
 			{
 				Code: "M04",
 				Name: i18n.String{
-					i18n.EN: "Exempt pursuant to Article 13 of the VAT code",
-					i18n.PT: "Isento artigo 13.º do CIVA",
+					i18n.EN: "Article 13 of the VAT code",
+					i18n.PT: "Artigo 13.º do CIVA",
 				},
 			},
 			{
 				Code: "M05",
 				Name: i18n.String{
-					i18n.EN: "Exempt pursuant to Article 14 of the VAT code",
-					i18n.PT: "Isento artigo 14.º do CIVA",
+					i18n.EN: "Article 14 of the VAT code",
+					i18n.PT: "Artigo 14.º do CIVA",
 				},
 			},
 			{
 				Code: "M06",
 				Name: i18n.String{
-					i18n.EN: "Exempt pursuant to Article 15 of the VAT code",
-					i18n.PT: "Isento artigo 15.º do CIVA",
+					i18n.EN: "Article 15 of the VAT code",
+					i18n.PT: "Artigo 15.º do CIVA",
 				},
 			},
 			{
 				Code: "M07",
 				Name: i18n.String{
-					i18n.EN: "Exempt pursuant to Article 9 of the VAT code",
-					i18n.PT: "Isento artigo 9.º do CIVA",
+					i18n.EN: "Article 9 of the VAT code",
+					i18n.PT: "Artigo 9.º do CIVA",
 				},
 			},
 			{
 				Code: "M09",
 				Name: i18n.String{
-					i18n.EN: "VAT - does not confer right to deduction / Article 62 paragraph b) of the VAT code",
-					i18n.PT: "IVA - não confere direito a dedução / Artigo 62.º alínea b) do CIVA",
+					i18n.EN: "Article 62 paragraph b) of the VAT code",
+					i18n.PT: "Artigo 62.º alínea b) do CIVA",
 				},
 			},
 			{
 				Code: "M10",
 				Name: i18n.String{
-					i18n.EN: "VAT - exemption scheme / Article 57 of the VAT code",
-					i18n.PT: "IVA - regime de isenção / Artigo 57.º do CIVA",
+					i18n.EN: "Article 57 of the VAT code",
+					i18n.PT: "Artigo 57.º do CIVA",
 				},
 			},
 			{
 				Code: "M11",
 				Name: i18n.String{
-					i18n.EN: "Special scheme for tobacco / Decree-Law No. 346/85 of 23rd August",
-					i18n.PT: "Regime particular do tabaco / Decreto-Lei n.º 346/85, de 23 de agosto",
+					i18n.EN: "Decree-Law No. 346/85 of 23rd August",
+					i18n.PT: "Decreto-Lei n.º 346/85, de 23 de agosto",
 				},
 			},
 			{
 				Code: "M12",
 				Name: i18n.String{
-					i18n.EN: "Margin scheme - Travel agencies / Decree-Law No. 221/85 of 3rd July",
-					i18n.PT: "Regime da margem de lucro - Agências de viagens / Decreto-Lei n.º 221/85, de 3 de julho",
+					i18n.EN: "Decree-Law No. 221/85 of 3rd July",
+					i18n.PT: "Decreto-Lei n.º 221/85, de 3 de julho",
 				},
 			},
 			{
 				Code: "M13",
 				Name: i18n.String{
-					i18n.EN: "Margin scheme - Second-hand goods / Decree-Law No. 199/96 of 18th October",
-					i18n.PT: "Regime da margem de lucro - Bens em segunda mão / Decreto-Lei n.º 199/96, de 18 de outubro",
+					i18n.EN: "Decree-Law No. 199/96 of 18th October",
+					i18n.PT: "Decreto-Lei n.º 199/96, de 18 de outubro",
 				},
 			},
 			{
 				Code: "M14",
 				Name: i18n.String{
-					i18n.EN: "Margin scheme - Works of art / Decree-Law No. 199/96 of 18th October",
-					i18n.PT: "Regime da margem de lucro - Objetos de arte / Decreto-Lei n.º 199/96, de 18 de outubro",
+					i18n.EN: "Decree-Law No. 199/96 of 18th October",
+					i18n.PT: "Decreto-Lei n.º 199/96, de 18 de outubro",
 				},
 			},
 			{
 				Code: "M15",
 				Name: i18n.String{
-					i18n.EN: "Margin scheme - Collector's items and antiques / Decree-Law No. 199/96 of 18th October",
-					i18n.PT: "Regime da margem de lucro - Objetos de coleção e antiguidades / Decreto-Lei n.º 199/96, de 18 de outubro",
+					i18n.EN: "Decree-Law No. 199/96 of 18th October",
+					i18n.PT: "Decreto-Lei n.º 199/96, de 18 de outubro",
 				},
 			},
 			{
 				Code: "M16",
 				Name: i18n.String{
-					i18n.EN: "Exempt pursuant to Article 14 of the RITI",
-					i18n.PT: "Isento artigo 14.º do RITI",
+					i18n.EN: "Article 14 of the RITI",
+					i18n.PT: "Artigo 14.º do RITI",
 				},
 			},
 			{
@@ -443,92 +480,92 @@ var extensions = []*cbc.Definition{
 			{
 				Code: "M20",
 				Name: i18n.String{
-					i18n.EN: "VAT - flat-rate scheme / Article 59-D No. 2 of the VAT code",
-					i18n.PT: "IVA - regime forfetário / Artigo 59.º-D n.º 2 do CIVA",
+					i18n.EN: "Article 59-D No. 2 of the VAT code",
+					i18n.PT: "Artigo 59.º-D n.º 2 do CIVA",
 				},
 			},
 			{
 				Code: "M21",
 				Name: i18n.String{
-					i18n.EN: "VAT - does not confer right to deduction (or similar) / Article 72 No. 4 of the VAT code",
-					i18n.PT: "IVA - não confere direito à dedução (ou expressão similar) / Artigo 72.º n.º 4 do CIVA",
+					i18n.EN: "Article 72 No. 4 of the VAT code",
+					i18n.PT: "Artigo 72.º n.º 4 do CIVA",
 				},
 			},
 			{
 				Code: "M25",
 				Name: i18n.String{
-					i18n.EN: "Consignment goods / Article 38 No. 1 paragraph a) of the VAT code",
-					i18n.PT: "Mercadoria à consignação / Artigo 38.º n.º 1 alínea a) do CIVA",
+					i18n.EN: "Article 38 No. 1 paragraph a) of the VAT code",
+					i18n.PT: "Artigo 38.º n.º 1 alínea a) do CIVA",
 				},
 			},
 			{
 				Code: "M26",
 				Name: i18n.String{
-					i18n.EN: "VAT exemption with right to deduction in food basket / Law No. 17/2023 of 14th April",
-					i18n.PT: "Isenção de IVA com direito à dedução no cabaz alimentar / Lei n.º 17/2023, de 14 de abril",
+					i18n.EN: "Law No. 17/2023 of 14th April",
+					i18n.PT: "Lei n.º 17/2023, de 14 de abril",
 				},
 			},
 			{
 				Code: "M30",
 				Name: i18n.String{
-					i18n.EN: "VAT - reverse charge / Article 2 No. 1 paragraph i) of the VAT code",
-					i18n.PT: "IVA - autoliquidação / Artigo 2.º n.º 1 alínea i) do CIVA",
+					i18n.EN: "Article 2 No. 1 paragraph i) of the VAT code",
+					i18n.PT: "Artigo 2.º n.º 1 alínea i) do CIVA",
 				},
 			},
 			{
 				Code: "M31",
 				Name: i18n.String{
-					i18n.EN: "VAT - reverse charge / Article 2 No. 1 paragraph j) of the VAT code",
-					i18n.PT: "IVA - autoliquidação / Artigo 2.º n.º 1 alínea j) do CIVA",
+					i18n.EN: "Article 2 No. 1 paragraph j) of the VAT code",
+					i18n.PT: "Artigo 2.º n.º 1 alínea j) do CIVA",
 				},
 			},
 			{
 				Code: "M32",
 				Name: i18n.String{
-					i18n.EN: "VAT - reverse charge / Article 2 No. 1 paragraph l) of the VAT code",
-					i18n.PT: "IVA - autoliquidação / Artigo 2.º n.º 1 alínea l) do CIVA",
+					i18n.EN: "Article 2 No. 1 paragraph l) of the VAT code",
+					i18n.PT: "Artigo 2.º n.º 1 alínea l) do CIVA",
 				},
 			},
 			{
 				Code: "M33",
 				Name: i18n.String{
-					i18n.EN: "VAT - reverse charge / Article 2 No. 1 paragraph m) of the VAT code",
-					i18n.PT: "IVA - autoliquidação / Artigo 2.º n.º 1 alínea m) do CIVA",
+					i18n.EN: "Article 2 No. 1 paragraph m) of the VAT code",
+					i18n.PT: "Artigo 2.º n.º 1 alínea m) do CIVA",
 				},
 			},
 			{
 				Code: "M34",
 				Name: i18n.String{
-					i18n.EN: "VAT - reverse charge / Article 2 No. 1 paragraph n) of the VAT code",
-					i18n.PT: "IVA - autoliquidação / Artigo 2.º n.º 1 alínea n) do CIVA",
+					i18n.EN: "Article 2 No. 1 paragraph n) of the VAT code",
+					i18n.PT: "Artigo 2.º n.º 1 alínea n) do CIVA",
 				},
 			},
 			{
 				Code: "M40",
 				Name: i18n.String{
-					i18n.EN: "VAT - reverse charge / Article 6 No. 6 paragraph a) of the VAT code, to the contrary",
-					i18n.PT: "IVA - autoliquidação / Artigo 6.º n.º 6 alínea a) do CIVA, a contrário",
+					i18n.EN: "Article 6 No. 6 paragraph a) of the VAT code, to the contrary",
+					i18n.PT: "Artigo 6.º n.º 6 alínea a) do CIVA, a contrário",
 				},
 			},
 			{
 				Code: "M41",
 				Name: i18n.String{
-					i18n.EN: "VAT - reverse charge / Article 8 No. 3 of the RITI",
-					i18n.PT: "IVA - autoliquidação / Artigo 8.º n.º 3 do RITI",
+					i18n.EN: "Article 8 No. 3 of the RITI",
+					i18n.PT: "Artigo 8.º n.º 3 do RITI",
 				},
 			},
 			{
 				Code: "M42",
 				Name: i18n.String{
-					i18n.EN: "VAT - reverse charge / Decree-Law No. 21/2007 of 29th January",
-					i18n.PT: "IVA - autoliquidação / Decreto-Lei n.º 21/2007, de 29 de janeiro",
+					i18n.EN: "Decree-Law No. 21/2007 of 29th January",
+					i18n.PT: "Decreto-Lei n.º 21/2007, de 29 de janeiro",
 				},
 			},
 			{
 				Code: "M43",
 				Name: i18n.String{
-					i18n.EN: "VAT - reverse charge / Decree-Law No. 362/99 of 16th September",
-					i18n.PT: "IVA - autoliquidação / Decreto-Lei n.º 362/99, de 16 de setembro",
+					i18n.EN: "Decree-Law No. 362/99 of 16th September",
+					i18n.PT: "Decreto-Lei n.º 362/99, de 16 de setembro",
 				},
 			},
 			{